Task 15 - Addressed review comments: MVVM Design Pattern - MVC design pattern is replaced with MVVM - Separate ActivityIndicatorView class is created - Unit testing is done after implementing MVVM design pattern - Naming convention for files/ folders is followed as per MVVM design pattern
